+++ pages/index.tsx
@@ -1,8 +1,4 @@
 import type { NextPage } from 'next';
-import dotenv from "dotenv";
-
-dotenv.config();
-console.log("너 엔브자나", process.env.REST_API_KEY);
 
 const Home: NextPage = () => <div />;
