+++ shared/index.d.ts
@@ -2,10 +2,3 @@ declare module '*.png';
 declare module '*.jpg';
 declare module '*.jpeg';
 declare module '*.gif';
-
-declare global {
-	interface window {
-		kakao: any;
-	}
-	const kakao: any;
-}
